hooks package refactored
