+++ src/main/java/com/configcat/ConfigurationProvider.java
@@ -4,6 +4,8 @@ import java.io.Closeable;
 import java.util.Collection;
 import java.util.List;
 import java.util.Map;
+
+import com.configcat.hooks.ConfigCatHooks;
 import java9.util.concurrent.CompletableFuture;
 
 /**
