+++ src/main/java/com/configcat/hooks/ConfigCatHooks.java
@@ -1,9 +1,11 @@
-package com.configcat;
+package com.configcat.hooks;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
+import com.configcat.EvaluationDetails;
+import com.configcat.Setting;
 import java9.util.function.Consumer;
 
 public class ConfigCatHooks {
