+++ src/test/java/com/configcat/ConfigFetcherTest.java
@@ -1,5 +1,6 @@
 package com.configcat;
 
+import com.configcat.hooks.ConfigCatHooks;
 import com.configcat.log.ConfigCatLogger;
 import com.configcat.log.LogLevel;
 import com.google.gson.Gson;
